embedding_subsystem/requirements.txt
huggingface-hub
ipykernel
ipywidgets
langchain
minio
psycopg2-binary
python-dotenv
ray[train,data]
sentence_transformers
virtualenv